backend/requirements.txt
flask
python-dotenv
flask-cors
openai
gunicorn